// src/constants/enum.ts
export enum EMenu {
  HOME = "home",
  CATEGORIES = "categories",
  SUB_CATEGORIES = "sub_categories",
  PRODUCTS = "products",
  FEATURED = "featured",
  BRANDS = "brands",
  POSTS = "posts",
  SLIDE_SMALL = "slide-small",
  ACCESSORIES = "accessories",
  GIFTS = "gifts",
  BANNERS = "banners",
  TOP_PRODUCTS = "top-products",
  USERS = "users",
  SETTINGS = "settings",
}
export enum SpecificationType {
  CPU = "CPU",
  VGA = "VGA",
}

export enum EUserRoles {
  ADMIN = "admin",
  USER = "user",
  STAFF = "staff",
}